refactor: Move nested networking package to root level api package
