+++ cardpresent/src/main/java/com/staxpayments/api/PaginatedData.kt
@@ -1,3 +1,3 @@
-package com.staxpayments.sdk.networking
+package com.staxpayments.api
 
 internal class PaginatedData<T>(val data: List<T>)
