+++ cardpresent/src/main/java/com/staxpayments/api/StaxApi.kt
@@ -1,19 +1,19 @@
-package com.staxpayments.sdk.networking
+package com.staxpayments.api
 
 import com.staxpayments.sdk.Environment
 import com.staxpayments.sdk.Environment.DEV
 import com.staxpayments.sdk.Environment.LIVE
 import com.staxpayments.sdk.Environment.QA
 import com.staxpayments.sdk.data.Amount
-import com.staxpayments.sdk.data.models.ChargeRequest
-import com.staxpayments.sdk.data.models.Customer
-import com.staxpayments.sdk.data.models.Invoice
-import com.staxpayments.sdk.data.models.Merchant
-import com.staxpayments.sdk.data.models.MobileReaderDetails
-import com.staxpayments.sdk.data.models.PaymentMethod
-import com.staxpayments.sdk.data.models.Self
-import com.staxpayments.sdk.data.models.Transaction
 import com.google.gson.FieldNamingPolicy
+import com.staxpayments.api.models.ChargeRequest
+import com.staxpayments.api.models.Customer
+import com.staxpayments.api.models.Invoice
+import com.staxpayments.api.models.Merchant
+import com.staxpayments.api.models.MobileReaderDetails
+import com.staxpayments.api.models.PaymentMethod
+import com.staxpayments.api.models.Self
+import com.staxpayments.api.models.Transaction
 import io.ktor.client.HttpClient
 import io.ktor.client.call.NoTransformationFoundException
 import io.ktor.client.call.body
