+++ cardpresent/src/main/java/com/staxpayments/api/models/CatalogItem.kt
@@ -1,4 +1,4 @@
-package com.staxpayments.sdk.data.models
+package com.staxpayments.api.models
 
 open class CatalogItem : Model {
     override var id: String? = null
