+++ cardpresent/src/main/java/com/staxpayments/api/models/ChargeRequest.kt
@@ -1,4 +1,4 @@
-package com.staxpayments.sdk.data.models
+package com.staxpayments.api.models
 
 import com.squareup.moshi.Json
