+++ cardpresent/src/main/java/com/staxpayments/api/models/Transaction.kt
@@ -1,4 +1,4 @@
-package com.staxpayments.sdk.data.models
+package com.staxpayments.api.models
 
 /**
  * A Stax Transaction
