+++ cardpresent/src/main/java/com/staxpayments/sdk/data/LogTransactionResponse.kt
@@ -1,6 +1,6 @@
 package com.staxpayments.sdk.data
 
-import com.staxpayments.sdk.data.models.Transaction
+import com.staxpayments.api.models.Transaction
 
 class LogTransactionResponse {
     val transaction: Transaction? = null
