+++ cardpresent/src/main/java/com/staxpayments/sdk/data/TransactionResult.kt
@@ -1,10 +1,10 @@
 package com.staxpayments.sdk.data
 
 import com.staxpayments.android.chipdna.ChipDnaDriver
-import com.staxpayments.sdk.data.models.Customer
-import com.staxpayments.sdk.data.models.Invoice
-import com.staxpayments.sdk.data.models.PaymentMethod
-import com.staxpayments.sdk.data.models.Transaction
+import com.staxpayments.api.models.Customer
+import com.staxpayments.api.models.Invoice
+import com.staxpayments.api.models.PaymentMethod
+import com.staxpayments.api.models.Transaction
 
 /**
  * Represents a request for a mobile reader payment
