+++ cardpresent/src/main/java/com/staxpayments/sdk/usecase/CancelCurrentTransaction.kt
@@ -1,7 +1,7 @@
 package com.staxpayments.sdk.usecase
 
 import com.staxpayments.exceptions.StaxException
-import com.staxpayments.sdk.data.repository.MobileReaderDriverRepository
+import com.staxpayments.api.repository.MobileReaderDriverRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlin.coroutines.CoroutineContext
