+++ cardpresent/src/main/java/com/staxpayments/sdk/usecase/DisconnectMobileReader.kt
@@ -3,7 +3,7 @@ package com.staxpayments.sdk.usecase
 import com.staxpayments.exceptions.DisconnectMobileReaderException
 import com.staxpayments.exceptions.StaxException
 import com.staxpayments.sdk.data.MobileReader
-import com.staxpayments.sdk.data.repository.MobileReaderDriverRepository
+import com.staxpayments.api.repository.MobileReaderDriverRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlin.coroutines.CoroutineContext
