+++ cardpresent/src/main/java/com/staxpayments/sdk/usecase/TakeMobileReaderPayment.kt
@@ -9,14 +9,14 @@ import com.staxpayments.sdk.UserNotificationListener
 import com.staxpayments.sdk.data.MobileReaderDriver
 import com.staxpayments.sdk.data.TransactionRequest
 import com.staxpayments.sdk.data.TransactionResult
-import com.staxpayments.sdk.data.models.Customer
-import com.staxpayments.sdk.data.models.Invoice
-import com.staxpayments.sdk.data.models.Transaction
-import com.staxpayments.sdk.data.repository.CustomerRepository
-import com.staxpayments.sdk.data.repository.InvoiceRepository
-import com.staxpayments.sdk.data.repository.MobileReaderDriverRepository
-import com.staxpayments.sdk.data.repository.PaymentMethodRepository
-import com.staxpayments.sdk.data.repository.TransactionRepository
+import com.staxpayments.api.models.Customer
+import com.staxpayments.api.models.Invoice
+import com.staxpayments.api.models.Transaction
+import com.staxpayments.api.repository.CustomerRepository
+import com.staxpayments.api.repository.InvoiceRepository
+import com.staxpayments.api.repository.MobileReaderDriverRepository
+import com.staxpayments.api.repository.PaymentMethodRepository
+import com.staxpayments.api.repository.TransactionRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
 import kotlin.coroutines.CoroutineContext
