+++ cardpresent/src/main/java/com/staxpayments/sdk/usecase/VoidTransaction.kt
@@ -1,8 +1,8 @@
 package com.staxpayments.sdk.usecase
 
 import com.staxpayments.exceptions.StaxException
-import com.staxpayments.sdk.data.models.Transaction
-import com.staxpayments.sdk.networking.StaxApi
+import com.staxpayments.api.models.Transaction
+import com.staxpayments.api.StaxApi
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
 import kotlin.coroutines.CoroutineContext
